+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.8.0
 - Add support for logical types. Currently this requires using the
-  `avro-salsify-fork` gem for logical support with Ruby.
+  `avro-salsify-fork` gem for logical types support with Ruby.
 
 ## v0.7.1
 - Raise a more descriptive error when attempting to generate a model for a